Add resource UID translator for resource permissions
