+++ pkg/services/accesscontrol/ossaccesscontrol/service_account.go
@@ -41,8 +41,9 @@ func ProvideServiceAccountPermissions(
 	teamService team.Service, userService user.Service, actionSetService resourcepermissions.ActionSetService,
 ) (*ServiceAccountPermissionsService, error) {
 	options := resourcepermissions.Options{
-		Resource:          "serviceaccounts",
-		ResourceAttribute: "id",
+		Resource:           "serviceaccounts",
+		ResourceAttribute:  "id",
+		ResourceTranslator: serviceaccounts.UIDToIDHandler(serviceAccountRetrieverService),
 		ResourceValidator: func(ctx context.Context, orgID int64, resourceID string) error {
 			ctx, span := tracer.Start(ctx, "accesscontrol.ossaccesscontrol.ProvideServiceAccountPermissions.ResourceValidator")
 			defer span.End()
