+++ pkg/services/serviceaccounts/api/api.go
@@ -31,30 +31,6 @@ type ServiceAccountsAPI struct {
 	isExternalSAEnabled  bool
 }
 
-func MiddlewareServiceAccountUIDResolver(saService serviceaccounts.Service, paramName string) web.Handler {
-	return func(c *contextmodel.ReqContext) {
-		// Get service account id from request
-		saUID := web.Params(c.Req)[paramName]
-		// if saID is empty or is an integer, we assume it's a service account id and we don't need to resolve it
-		_, err := strconv.ParseInt(saUID, 10, 64)
-		if saUID == "" || err == nil {
-			return
-		}
-
-		serviceAccount, err := saService.RetrieveServiceAccount(c.Req.Context(), &serviceaccounts.GetServiceAccountQuery{
-			OrgID: c.SignedInUser.GetOrgID(),
-			UID:   saUID,
-		})
-		if err == nil {
-			gotParams := web.Params(c.Req)
-			gotParams[paramName] = strconv.FormatInt(serviceAccount.Id, 10)
-			web.SetURLParams(c.Req, gotParams)
-		} else {
-			c.JsonApiErr(http.StatusNotFound, "Not found", nil)
-		}
-	}
-}
-
 func NewServiceAccountsAPI(
 	cfg *setting.Cfg,
 	service serviceaccounts.Service,
@@ -79,7 +55,7 @@ func NewServiceAccountsAPI(
 
 func (api *ServiceAccountsAPI) RegisterAPIEndpoints() {
 	auth := accesscontrol.Middleware(api.accesscontrol)
-	saUIDResolver := MiddlewareServiceAccountUIDResolver(api.service, ":serviceAccountId")
+	saUIDResolver := serviceaccounts.MiddlewareServiceAccountUIDResolver(api.service, ":serviceAccountId")
 	api.RouterRegister.Group("/api/serviceaccounts", func(serviceAccountsRoute routing.RouteRegister) {
 		serviceAccountsRoute.Get("/search", auth(accesscontrol.EvalPermission(serviceaccounts.ActionRead)), routing.Wrap(api.SearchOrgServiceAccountsWithPaging))
 		serviceAccountsRoute.Post("/", auth(accesscontrol.EvalPermission(serviceaccounts.ActionCreate)), routing.Wrap(api.CreateServiceAccount))
