+++ pkg/services/serviceaccounts/retriever/retriever.go
@@ -13,13 +13,6 @@ import (
 	"github.com/grafana/grafana/pkg/services/user"
 )
 
-// ServiceAccountRetriever is the service that retrieves service accounts.
-// At the time of writing, this service is only used by the service accounts permissions service
-// to avoid cyclic dependency between the ServiceAccountService and the ServiceAccountPermissionsService
-type ServiceAccountRetriever interface {
-	RetrieveServiceAccount(ctx context.Context, query *serviceaccounts.GetServiceAccountQuery) (*serviceaccounts.ServiceAccountProfileDTO, error)
-}
-
 // ServiceAccountRetriever is the service that manages service accounts.
 type Service struct {
 	store  *database.ServiceAccountsStoreImpl
