+++ pkg/services/serviceaccounts/serviceaccounts.go
@@ -2,8 +2,12 @@ package serviceaccounts
 
 import (
 	"context"
+	"net/http"
+	"strconv"
 
 	"github.com/grafana/grafana/pkg/services/apikey"
+	contextmodel "github.com/grafana/grafana/pkg/services/contexthandler/model"
+	"github.com/grafana/grafana/pkg/web"
 )
 
 /*
@@ -13,11 +17,18 @@ Service accounts are used to authenticate API requests. They are not users and
 do not have a password.
 */
 
+// ServiceAccountRetriever is the service that retrieves service accounts.
+// At the time of writing, this service is only used by the service accounts permissions service
+// to avoid cyclic dependency between the ServiceAccountService and the ServiceAccountPermissionsService
+type ServiceAccountRetriever interface {
+	RetrieveServiceAccount(ctx context.Context, query *GetServiceAccountQuery) (*ServiceAccountProfileDTO, error)
+}
+
 //go:generate mockery --name Service --structname MockServiceAccountService --output tests --outpkg tests --filename mocks.go
 type Service interface {
+	ServiceAccountRetriever
 	CreateServiceAccount(ctx context.Context, orgID int64, saForm *CreateServiceAccountForm) (*ServiceAccountDTO, error)
 	DeleteServiceAccount(ctx context.Context, orgID, serviceAccountID int64) error
-	RetrieveServiceAccount(ctx context.Context, query *GetServiceAccountQuery) (*ServiceAccountProfileDTO, error)
 	RetrieveServiceAccountIdByName(ctx context.Context, orgID int64, name string) (int64, error)
 	SearchOrgServiceAccounts(ctx context.Context, query *SearchOrgServiceAccountsQuery) (*SearchOrgServiceAccountsResult, error)
 	EnableServiceAccount(ctx context.Context, orgID, serviceAccountID int64, enable bool) error
@@ -45,4 +56,36 @@ type ExtSvcAccountsService interface {
 	RemoveExtSvcAccount(ctx context.Context, orgID int64, extSvcSlug string) error
 	// RetrieveExtSvcAccount fetches an external service account by ID
 	RetrieveExtSvcAccount(ctx context.Context, orgID, saID int64) (*ExtSvcAccount, error)
+}
+
+func UIDToIDHandler(saService ServiceAccountRetriever) func(ctx context.Context, orgID int64, resourceID string) (string, error) {
+	return func(ctx context.Context, orgID int64, resourceID string) (string, error) {
+		// if saID is empty or is an integer, we assume it's a team id and we don't need to resolve it
+		_, err := strconv.ParseInt(resourceID, 10, 64)
+		if resourceID == "" || err == nil {
+			return resourceID, nil
+		}
+		serviceAccount, err := saService.RetrieveServiceAccount(ctx, &GetServiceAccountQuery{
+			OrgID: orgID,
+			UID:   resourceID,
+		})
+		return strconv.FormatInt(serviceAccount.Id, 10), err
+	}
+}
+
+func MiddlewareServiceAccountUIDResolver(saService Service, paramName string) web.Handler {
+	handler := UIDToIDHandler(saService)
+
+	return func(c *contextmodel.ReqContext) {
+		// Get sa id from request, fetch team and replace teamId with team id
+		saUID := web.Params(c.Req)[paramName]
+		id, err := handler(c.Req.Context(), c.SignedInUser.GetOrgID(), saUID)
+		if err == nil {
+			gotParams := web.Params(c.Req)
+			gotParams[paramName] = id
+			web.SetURLParams(c.Req, gotParams)
+		} else {
+			c.JsonApiErr(http.StatusNotFound, "Not found", nil)
+		}
+	}
 }
